+++ cpp/src/hash/hashing.cu
@@ -53,17 +53,18 @@ template <template <typename> class hash_function,
          typename size_type>
 struct row_hasher
 {
-  row_hasher(gdf_table<size_type> const & table_to_hash)
-    : the_table{table_to_hash}
+  row_hasher(gdf_table<size_type> const & table_to_hash, hash_value_type *initial_hash_values)
+    : the_table{table_to_hash}, initial_hash_values(initial_hash_values)
   {}
 
   __device__
   hash_value_type operator()(size_type row_index) const
   {
-    return the_table.template hash_row<hash_function>(row_index);
+    return the_table.template hash_row<hash_function>(row_index, 0, initial_hash_values);
   }
 
   gdf_table<size_type> const & the_table;
+  hash_value_type* initial_hash_values{nullptr};
 };
 
 
@@ -75,12 +76,17 @@ struct row_hasher
  * @Param num_cols The number of columns in the input set
  * @Param input The list of columns whose rows will be hashed
  * @Param hash The hash function to use
+ * @Param initial_hash_values Initial hash values to combine with column hash values
  * @Param output The hash value of each row of the input
  * 
  * @Returns   
  */
 /* ----------------------------------------------------------------------------*/
-gdf_error gdf_hash(int num_cols, gdf_column **input, gdf_hash_func hash, gdf_column *output)
+gdf_error gdf_hash(int num_cols,
+                   gdf_column **input,
+                   gdf_hash_func hash,
+                   hash_value_type *initial_hash_values,
+                   gdf_column *output)
 {
   // Ensure inputs aren't null
   if((0 == num_cols)
@@ -128,9 +134,9 @@ gdf_error gdf_hash(int num_cols, gdf_column **input, gdf_hash_func hash, gdf_col
     case GDF_HASH_MURMUR3:
       {
         thrust::tabulate(rmm::exec_policy()->on(0),
-                        row_hash_values, 
+                         row_hash_values,
                          row_hash_values + num_rows, 
-                         row_hasher<MurmurHash3_32,size_type>(*input_table));
+                         row_hasher<MurmurHash3_32,size_type>(*input_table, initial_hash_values));
         break;
       }
     case GDF_HASH_IDENTITY:
@@ -138,7 +144,7 @@ gdf_error gdf_hash(int num_cols, gdf_column **input, gdf_hash_func hash, gdf_col
         thrust::tabulate(rmm::exec_policy()->on(0),
                          row_hash_values, 
                          row_hash_values + num_rows, 
-                         row_hasher<IdentityHash,size_type>(*input_table));
+                         row_hasher<IdentityHash,size_type>(*input_table, initial_hash_values));
         break;
       }
     default: